requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.5.0
    # via httpcore
asgiref==3.4.1
    # via django
astor==0.8.1
    # via flake8-simplify
attrs==21.4.0
    # via
    #   flake8-bugbear
    #   flake8-eradicate
    #   flake8-multiline-containers
    #   jsonschema
    #   pytest
certifi==2021.10.8
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.10
    # via
    #   httpx
    #   requests
click==8.0.3
    # via pip-tools
cognitive-complexity==1.2.0
    # via flake8-cognitive-complexity
coreapi==2.3.3
    # via
    #   django-rest-swagger
    #   openapi-codec
coreschema==0.0.4
    # via coreapi
cryptography==36.0.1
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
django==4.0.1
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-filter
    #   django-rest-auth
    #   djangorestframework
    #   drf-spectacular
django-allauth==0.47.0
    # via django-rest-auth
django-filter==21.1
    # via -r requirements.in
django-rest-auth[with_social]==0.9.5
    # via -r requirements.in
django-rest-swagger==2.2.0
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   django-rest-auth
    #   django-rest-swagger
    #   drf-spectacular
drf-spectacular==0.21.1
    # via -r requirements.in
eradicate==2.0.0
    # via flake8-eradicate
execnet==1.9.0
    # via pytest-xdist
faker==11.3.0
    # via mixer
flake8==3.9.2
    # via
    #   -r requirements.in
    #   flake8-absolute-import
    #   flake8-bugbear
    #   flake8-commas
    #   flake8-django
    #   flake8-eradicate
    #   flake8-isort
    #   flake8-multiline-containers
    #   flake8-mutable
    #   flake8-pep3101
    #   flake8-print
    #   flake8-printf-formatting
    #   flake8-pytest
    #   flake8-quotes
    #   flake8-simplify
    #   flake8-use-fstring
    #   flake8-walrus
flake8-absolute-import==1.0.0.1
    # via -r requirements.in
flake8-bugbear==22.1.11
    # via -r requirements.in
flake8-cognitive-complexity==0.1.0
    # via -r requirements.in
flake8-commas==2.1.0
    # via -r requirements.in
flake8-django==1.1.2
    # via -r requirements.in
flake8-eradicate==1.2.0
    # via -r requirements.in
flake8-fixme==1.1.1
    # via -r requirements.in
flake8-isort==4.1.1
    # via -r requirements.in
flake8-mock==0.3
    # via -r requirements.in
flake8-multiline-containers==0.0.18
    # via -r requirements.in
flake8-mutable==1.2.0
    # via -r requirements.in
flake8-pep3101==1.3.0
    # via -r requirements.in
flake8-pie==0.15.0
    # via -r requirements.in
flake8-plugin-utils==1.3.2
    # via flake8-pytest-style
flake8-print==4.0.0
    # via -r requirements.in
flake8-printf-formatting==1.1.2
    # via -r requirements.in
flake8-pytest==1.3
    # via -r requirements.in
flake8-pytest-style==1.6.0
    # via -r requirements.in
flake8-quotes==3.3.1
    # via -r requirements.in
flake8-simplify==0.14.5
    # via -r requirements.in
flake8-todo==0.7
    # via -r requirements.in
flake8-use-fstring==1.3
    # via -r requirements.in
flake8-variables-names==0.0.4
    # via -r requirements.in
flake8-walrus==1.1.0
    # via -r requirements.in
freezegun==1.1.0
    # via
    #   -r requirements.in
    #   pytest-freezegun
h11==0.12.0
    # via httpcore
httpcore==0.14.4
    # via httpx
httpx==0.21.3
    # via pytest-httpx
idna==3.3
    # via
    #   anyio
    #   requests
    #   rfc3986
importlib-metadata==4.10.1
    # via pytest-randomly
inflection==0.5.1
    # via drf-spectacular
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via flake8-isort
itypes==1.2.0
    # via coreapi
jinja2==3.0.3
    # via coreschema
jsonschema==4.4.0
    # via drf-spectacular
markupsafe==2.0.1
    # via jinja2
mccabe==0.6.1
    # via flake8
mixer==7.2.0
    # via -r requirements.in
oauthlib==3.1.1
    # via requests-oauthlib
openapi-codec==1.3.2
    # via django-rest-swagger
packaging==21.3
    # via pytest
pep517==0.12.0
    # via pip-tools
pip-tools==6.4.0
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via
    #   pytest
    #   pytest-forked
pycodestyle==2.7.0
    # via
    #   flake8
    #   flake8-print
    #   flake8-todo
pycparser==2.21
    # via cffi
pyflakes==2.3.1
    # via flake8
pyjwt[crypto]==2.3.0
    # via django-allauth
pyparsing==3.0.6
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
pytest==6.2.5
    # via
    #   -r requirements.in
    #   pytest-deadfixtures
    #   pytest-django
    #   pytest-env
    #   pytest-forked
    #   pytest-freezegun
    #   pytest-httpx
    #   pytest-mock
    #   pytest-randomly
    #   pytest-xdist
pytest-deadfixtures==2.2.1
    # via -r requirements.in
pytest-django==4.5.2
    # via -r requirements.in
pytest-env==0.6.2
    # via -r requirements.in
pytest-forked==1.4.0
    # via pytest-xdist
pytest-freezegun==0.4.2
    # via -r requirements.in
pytest-httpx==0.18.0
    # via -r requirements.in
pytest-mock==3.6.1
    # via -r requirements.in
pytest-randomly==3.11.0
    # via -r requirements.in
pytest-xdist==2.5.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   faker
    #   freezegun
python3-openid==3.2.0
    # via django-allauth
pytz==2021.3
    # via djangorestframework
pyyaml==6.0
    # via drf-spectacular
requests==2.27.1
    # via
    #   coreapi
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via django-allauth
rfc3986[idna2008]==1.5.0
    # via httpx
simplejson==3.17.6
    # via django-rest-swagger
six==1.16.0
    # via
    #   django-rest-auth
    #   flake8-print
    #   python-dateutil
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlparse==0.4.2
    # via django
testfixtures==6.18.3
    # via flake8-isort
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via pytest
tomli==2.0.0
    # via pep517
typing-extensions==4.0.1
    # via flake8-pie
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-spectacular
urllib3==1.26.8
    # via requests
wheel==0.37.1
    # via pip-tools
zipp==3.7.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
